+++ terraform/infra/modules/ecs/service_discovery.tf
@@ -30,8 +30,4 @@ resource "aws_service_discovery_service" "imap" {
 
     routing_policy = "MULTIVALUE"
   }
-
-  health_check_custom_config {
-    failure_threshold = 1
-  }
 }